requirements.txt
numpy
pandas
matplotlib
tensorflow-gpu
opencv-python
opencv-contrib-python
qrcode
pillow